refactor app.py and add test coverage there
