+++ test_data.py
@@ -1,0 +1,38 @@
+test_config = {
+    "1" : {
+    "tempMax": 60,
+    "tempMin": 40,
+    "humidityMax": 70,
+    "humidityMin": 60,
+    },
+    "2" : {
+    "tempMax": 80,
+    "tempMin": 50,
+    "humidityMax": 90,
+    "humidityMin": 80,
+    },
+    "3" : {
+    "tempMax": 80,
+    "tempMin": 50,
+    "humidityMax": 90,
+    "humidityMin": 80,
+    },
+    "4" : {
+    "tempMax": 80,
+    "tempMin": 50,
+    "humidityMax": 90,
+    "humidityMin": 80,
+    },
+    "5" : {
+    "tempMax": 80,
+    "tempMin": 50,
+    "humidityMax": 90,
+    "humidityMin": 80,
+    },
+    "6" : {
+    "tempMax": 80,
+    "tempMin": 50,
+    "humidityMax": 90,
+    "humidityMin": 80,
+    }
+}
